+++ index.js
@@ -104,3 +104,5 @@ export { default as abstractEquality } from './src/equality/abstractEquality';
 export { default as strictEquality } from './src/equality/strictEquality';
 export { default as sameValue } from './src/equality/sameValue';
 export { default as sameValueZero } from './src/equality/sameValueZero';
+
+export { default as addon } from './src/addon';