+++ src/addon.js
@@ -1,0 +1,6 @@
+import { superimpose } from 'object-agent';
+import * as typeEnforcer from '../index';
+
+export default (library) => {
+	superimpose(typeEnforcer, library, true);
+};